@@ -6,7 +6,7 @@ use Illuminate\Database\Eloquent\Model;
 
 class Post extends Model
 {
-    protected $fillable = [
+    protected array $fillable = [
         'slug',
         'title',
         'summary',
